requirements.txt
python-dotenv>=1.0.1
langchain>=0.2.14
langchain-openai>=0.2.3
langgraph>=0.2.16
langchain-community>=0.2.12
tavily-python>=0.3.10
matplotlib>=3.8.0
pandas>=2.2.2
